requirements.txt
openai
pandas
python-dotenv
numpy
fastapi[all]
uvicorn